Split SquareGraph out and make SongProgressGraph also able to take a list of Objects.
